requirements.txt
groq
python-dotenv
openai-whisper
sounddevice
soundfile
opencv-python
mediapipe==0.10.9
numpy
